+++ src/cid-util.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const mh = require('multihashes')
-
+const { Buffer } = require('buffer')
 var CIDUtil = {
   /**
    * Test if the given input is a valid CID object.
@@ -26,10 +26,10 @@ var CIDUtil = {
 
     if (other.version === 0) {
       if (other.codec !== 'dag-pb') {
-        return `codec must be 'dag-pb' for CIDv0`
+        return "codec must be 'dag-pb' for CIDv0"
       }
       if (other.multibaseName !== 'base58btc') {
-        return `multibaseName must be 'base58btc' for CIDv0`
+        return "multibaseName must be 'base58btc' for CIDv0"
       }
     }